LCD-44249 update boot2 package names
