+++ modules/util/client-extension-util-spring-boot2/src/main/java/com/liferay/client/extension/util/spring/boot2/BaseRestController.java
@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.client.extension.util.spring.boot;
+package com.liferay.client.extension.util.spring.boot2;
 
-import com.liferay.client.extension.util.spring.boot.service.BaseService;
+import com.liferay.client.extension.util.spring.boot2.service.BaseService;
 
 import org.springframework.web.bind.annotation.RestController;
