+++ modules/util/client-extension-util-spring-boot2/src/main/java/com/liferay/client/extension/util/spring/boot2/LiferayOAuth2ResourceServerEnableWebSecurity.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.client.extension.util.spring.boot;
+package com.liferay.client.extension.util.spring.boot2;
 
 import com.nimbusds.jose.JOSEObjectType;
 import com.nimbusds.jose.proc.DefaultJOSEObjectTypeVerifier;
