+++ modules/util/client-extension-util-spring-boot2/src/test/resources/com/liferay/client/extension/util/spring/boot2/LiferayOAuth2ResourceServerEnableWebSecurityTest.properties
